+++ extensions/obsidian/src/utils/interfaces.tsx
@@ -5,30 +5,40 @@ export interface FormValue {
   tags: string[];
 }
 
-export interface Preferences {
+export interface GlobalPreferences {
   vaultPath: string;
 }
 
-export interface NoteFormPreferences extends Preferences {
-  prefPath: string;
-  prefTag: string;
-  tags: string;
-  openOnCreate: boolean;
-  prefNoteName: string;
-  folderActions: string;
-}
-
-export interface SearchNotePreferences extends Preferences {
-  primaryAction: string;
-  excludedFolders: string;
+export interface QuickLookPreferences {
   removeYAML: boolean;
   removeLinks: boolean;
   removeLatex: boolean;
-  appendPrefix: string;
+}
+
+export interface AppendNotePreferences {
+  appendTemplate: string;
+  appendSelectedTemplate: string;
+}
+export interface NoteFormPreferences extends GlobalPreferences {
+  prefPath: string;
+  prefNoteName: string;
+  prefNoteContent: string;
+  fillFormWithDefaults: boolean;
+  prefTag: string;
+  tags: string;
+  openOnCreate: boolean;
+  folderActions: string;
+}
+
+export interface SearchNotePreferences extends GlobalPreferences, QuickLookPreferences, AppendNotePreferences {
+  primaryAction: string;
+  excludedFolders: string;
   showDetail: boolean;
   showMetadata: boolean;
   searchContent: boolean;
 }
+
+export interface RandomNotePreferences extends GlobalPreferences, QuickLookPreferences, AppendNotePreferences {}
 
 export interface Vault {
   name: string;
@@ -38,22 +48,27 @@ export interface Vault {
 
 export interface Note {
   title: string;
-  key: number;
   path: string;
+  tags: string[];
   content: string;
 }
 
-interface ObsidianJsonVault {
+interface ObsidianVaultJSON {
   path: string;
   ts: number;
   open: boolean;
 }
 
-export interface ObsidianJson {
-  vaults: Record<string, ObsidianJsonVault>;
+export interface ObsidianJSON {
+  vaults: Record<string, ObsidianVaultJSON>;
 }
 
 export interface ObsidianVaultsState {
   ready: boolean;
   vaults: Vault[];
+}
+
+export interface PinnedNotesJSON {
+  vaultPath: string;
+  pinnedNotes: string[];
 }