+++ extensions/craftdocs/CHANGELOG.md
@@ -1,5 +1,9 @@
 # Craftdocs Changelog
 
+## [Update] - 2022-07-12
+
+- Added a new `Daily Notes` command.
+
 ## [Update] - 2022-07-09
 
 - Updated icons in the list.